requirements.txt
opencv-python>=4.5.1
numpy>=1.19.2
Pillow>=8.1.0
boto3>=1.16.25

